Fixed pipeline

+++ embedded/embedded-firmware/src/firmware_sensors.cc
@@ -50,5 +50,5 @@ extern "C" {
 [[nodiscard]] float FirmwareSensors::getBatteryLevel() const {
   constexpr float kToPercentage = 10.0F;
   logVarId_t vbatid = logGetVarId("pm", "batteryLevel");
-  return logGetUint(vbatid) / kToPercentage;
+  return static_cast<float>(logGetUint(vbatid)) / kToPercentage;
 }
